Revert "Example 2 image (#24)"
This reverts commit e12f081e234ba13aa5a46f9576e650bb724adb30.
